Remove useless "Memento" complexity
