+++ app/Src/UseCases/Domain/Context/Model/Characteristic.php
@@ -7,10 +7,8 @@ namespace App\Src\UseCases\Domain\Context\Model;
 
 
 use App\Src\UseCases\Domain\Ports\CharacteristicsRepository;
-use App\Src\UseCases\Domain\Shared\Model\HasMemento;
-use App\Src\UseCases\Domain\Shared\Model\Memento;
 
-class Characteristic implements HasMemento
+class Characteristic
 {
     const FARMING_TYPE = 'farming';
     const CROPPING_SYSTEM = 'croppingSystem';
@@ -50,18 +48,38 @@ class Characteristic implements HasMemento
         app(CharacteristicsRepository::class)->save($this);
     }
 
-    public function memento(): Memento
-    {
-        return new CharacteristicMemento($this->id, $this->type, $this->title, $this->visible, $this->icon, $this->pageId, $this->wiki);
-    }
-
     public function id(): string
     {
         return $this->id;
     }
 
+    public function type(): string
+    {
+        return $this->type;
+    }
+
+    public function title(): string
+    {
+        return $this->title;
+    }
+
+    public function visible(): bool
+    {
+        return $this->visible;
+    }
+
+    public function icon(): ?string
+    {
+        return $this->icon;
+    }
+
     public function pageId(): ?int
     {
         return $this->pageId;
+    }
+
+    public function wiki(): string
+    {
+        return $this->wiki;
     }
 }
