--- app/Src/UseCases/Domain/Shared/Model/HasMemento.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace App\Src\UseCases\Domain\Shared\Model;
-
-
-interface HasMemento
-{
-    public function memento():Memento;
-}
