--- app/Src/UseCases/Domain/Shared/Model/Memento.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace App\Src\UseCases\Domain\Shared\Model;
-
-
-interface Memento
-{
-
-}
